release(matching, tag, region, restaurant)

See merge request s12-fintech-finance-sub1/S12P21C208!168

+++ backend/src/main/java/com/patriot/fourlipsclover/match/dto/response/LocalsProposalResponse.java
@@ -15,7 +15,7 @@ public class LocalsProposalResponse {
 
     private Integer proposalId;       // 기획서 고유 ID
     private Integer matchId;          // 연결된 매칭의 ID
-    private List<Long> restaurantIds; // 추천 식당들의 ID 목록 (필요시 이름이나 DTO로 확장 가능)
+    private List<Integer> restaurantIds; // 추천 식당들의 ID 목록 (필요시 이름이나 DTO로 확장 가능)
     private String recommendMenu;     // 추천 메뉴
     private String description;       // 기획서 상세 설명
